+++ contributor-plugins/vcs/vcs-plugin-api/src/main/kotlin/com/mreil/easy/vcs/VcsNone.kt
@@ -23,5 +23,7 @@ internal class VcsNone(
     override fun addAndCommit(
         paths: List<String>,
         message: String,
-    ): Provider<Boolean> = providers.provider { false }
+    ): Provider<Boolean> = providers.provider { true }
+
+    override fun tag(name: String): Provider<Boolean> = providers.provider { true }
 }